Break out prepare module, major test refactor
