+++ test/parse/index.js
@@ -9,7 +9,7 @@ describe('parse/index (parseAll)', () => {
     var opts = config.parseOptions(config.fixtureOpts);
     return parseAll(opts).then(allData => {
       expect(allData).to.be.an('object').and.to.have
-        .keys('data', 'pages', 'patterns');
+        .keys('data', 'pages', 'patterns', 'options');
     });
   });
 });
